fix(recipe): Fix a couple of bugs in the recipe
The recipe now computes direct sun hours regardless of whether the input wea has direct irradiance for a certain hour. This also fixes a bug that results from the fact that the defaults of the ConvertToBinary function changed.

Lastly, it removes the direct radiation output since it no longer has any meaning when all of the Wea values are the same.