fix(result): Remove unused/incorrect direct radiation result
I realized that we were still copying the grids_info.json into a radiation result folder even though we were no longer reporting such radiation results (they are inaccurate given the way that we change the Wea values to be all the same.

Also, we still had a disclaimer about the radiance modifier limitations in the Readme, which is no longer relevant.